Move Comments to its own folder
