+++ src/components/Comments/Comments.less
@@ -1,4 +1,4 @@
-@import "../styles/custom.less";
+@import "../../styles/custom.less";
 
 .Comments {
   max-width: 990px;
@@ -42,7 +42,7 @@
     }
 
     &__footer {
-      padding: 4px 0;
+      padding: 8px 0;
       color: #99aab5;
       line-height: 14px;
